consolidate common/mini-game into common/mini-board
